+++ base/src/main/scala/org/opentorah/store/Stores.scala
@@ -5,37 +5,37 @@ import org.opentorah.xml.Parser
 import zio.ZIO
 
 trait Stores {
+  // TODO add getBys and use it in resolve()!
+
   def findByName(name: String): Caching.Parser[Option[Store]]
+}
 
-  final def findByNameInStores(name: String):  Parser[Option[Store]] =
-    findByNameInStores(name, stores)
+object Stores {
 
-  def stores: Seq[Store]
+  trait NonTerminal extends Stores {
 
-  final def findByName[M](
-    fullName: String,
-    findByName: String => Caching.Parser[Option[M]],
-    allowedExtension: String,
-    assumeAllowedExtension: Boolean
-  ): Caching.Parser[Option[M]] = {
-    val (fileName: String, extension: Option[String]) = Files.nameAndExtension(fullName)
+    override def findByName(name: String): Caching.Parser[Option[Store]] =
+      findByNameAmongNonTerminalStores(name)
 
-    val name: Option[String] =
-      if (extension.isDefined && !extension.contains(allowedExtension))
-        if (assumeAllowedExtension) Some(fullName) else None
-      else
-        Some(fileName)
+    protected final def findByNameAmongNonTerminalStores(name: String): Caching.Parser[Option[Store.NonTerminal]] =
+      findByNameAmongStores(name, nonTerminalStores)
 
-    name.fold[Caching.Parser[Option[M]]](ZIO.none)(findByName)
+    protected def nonTerminalStores: Seq[Store.NonTerminal]
   }
 
-  // Note: Site calls this twice, and that is why it is not unfolded right here;
-  // it's time to start distinguishing terminal stores from non-terminal ones (containers) -
-  // or, rather, FindByNames, not Stores;
-  // see also acceptsIndexHtml below.
-  // Then maybe more rigidity can be introduced and the double call removed...
-  final def findByNameInStores(name: String, stores: Seq[Store]): Parser[Option[Store]] =
-    ZIO.succeed(stores.find(_.names.hasName(name)))
+  trait Terminal extends Stores {
+
+    override def findByName(name: String): Caching.Parser[Option[Store]] =
+      findByNameAmongTerminalStores(name)
+
+    protected final def findByNameAmongTerminalStores(name: String): Caching.Parser[Option[Store.Terminal]] =
+      findByNameWithExtension(
+        name = name,
+        findByName = name => findByNameAmongStores(name, terminalStores)
+      )
+
+    protected def terminalStores: Seq[Store.Terminal]
+  }
 
   /*
     Successful `resolve()` returns a 'path' - a sequence of `Stores`;
@@ -44,24 +44,44 @@ trait Stores {
 
     Not all `Stores` are read from XML - some are constructed -
     so `Store` does *not* extend `FromUrl.With`.
-  */
+   */
   def resolve(
     path: Seq[String],
+    current: Stores,
     acc: Store.Path
-  ): Caching.Parser[Option[Store.Path]] = {
-    if (isEndOfPath(path)) endOfPath(acc) else for {
-      nextOption <- findByName(path.head)
-      result <- nextOption match {
-        case Some(next: Stores) => next.resolve(path.tail, next +: acc)
-        case Some(next) if isEndOfPath(path.tail) => endOfPath(next +: acc)
-        case _ => ZIO.succeed(None) // Note: with the None pattern, I get "match may not be exhaustive" warning...
-      }
-    } yield result
+  ): Caching.Parser[Option[Store.Path]] =
+    if (Stores.isEndOfPath(path)) Stores.endOfPath(acc) else current.findByName(path.head) flatMap {
+      case Some(next: Stores) => resolve(path.tail, next, next +: acc)
+      case Some(next) if Stores.isEndOfPath(path.tail) => Stores.endOfPath(next +: acc)
+      case _ => ZIO.succeed(None) // Note: with the None pattern, I get "match may not be exhaustive" warning...
+    }
+
+  private def findByNameAmongStores[T <: Store](
+    name: String,
+    stores: Seq[T]
+  ): Parser[Option[T]] =
+    ZIO.succeed(stores.find(_.names.hasName(name)))
+
+  def findByNameWithExtension[M <: Store](
+    name: String,
+    findByName: String => Caching.Parser[Option[M]],
+    allowedExtension: String = "html",
+    assumeAllowedExtension: Boolean = false
+  ): Caching.Parser[Option[M]] = {
+    val (fileName: String, extension: Option[String]) = Files.nameAndExtension(name)
+
+    val result: Option[String] =
+      if (extension.isDefined && !extension.contains(allowedExtension))
+        if (assumeAllowedExtension) Some(name) else None
+      else
+        Some(fileName)
+
+    result.fold[Caching.Parser[Option[M]]](ZIO.none)(findByName)
   }
 
   // TODO this is too HTML-ly for the 'store' package and should be something like isDirectory/isTerminal
   // (or be handled by the caller) ...
-  private def isEndOfPath(path: Seq[String]): Boolean =
+  def isEndOfPath(path: Seq[String]): Boolean =
     path.isEmpty || (path == Seq("index.html")) || (path == Seq("index"))
 
   private def endOfPath(acc: Store.Path):zio.UIO[Option[Store.Path]] =